Rename the allocation API from v1alpha1 -> v1.
